refactor(decompose): split packages/measures/perf/electron-main-storm.ts into folder
