+++ packages/measures/perf/electron-main-storm.ts
@@ -32,25 +32,16 @@
  */
 
 import {
-    existsSync,
     mkdirSync,
     mkdtempSync,
     readFileSync,
     rmSync,
 } from 'node:fs'
 import { homedir, tmpdir } from 'node:os'
-import { dirname, join, resolve } from 'node:path'
-import { spawn, type ChildProcessWithoutNullStreams } from 'node:child_process'
-import { createRequire } from 'node:module'
-import { fileURLToPath } from 'node:url'
+import { dirname, join } from 'node:path'
+import type { ChildProcessWithoutNullStreams } from 'node:child_process'
 
-import { agentRuntime, configureAgentRuntime } from '@vt/agent-runtime'
-import {
-    createTerminalData,
-    type TerminalData,
-    type TerminalId,
-} from '@vt/agent-runtime/types'
-import { initGraphModel } from '@vt/graph-model'
+import { agentRuntime } from '@vt/agent-runtime'
 import { generateVaultOnDisk } from '@vt/perf-fixtures'
 
 import {
@@ -60,385 +51,10 @@ import {
     stopMainProcessProfileAndSave,
     type MainProcessCdpHandle,
 } from './_shared/main-process-cdp.ts'
-
-// ---------------------------------------------------------------------------
-// Args
-// ---------------------------------------------------------------------------
-
-interface Args {
-    readonly agents: number
-    readonly nodesPerAgent: number
-    readonly vaultSeedNodeCount: number
-    readonly perAgentTimeoutMs: number
-    readonly bootTimeoutMs: number
-    readonly settleAfterStormMs: number
-    readonly outPath: string | null
-    readonly keepArtifacts: boolean
-}
-
-function parseArgs(argv: readonly string[]): Args {
-    const defaults: Args = {
-        agents: 5,
-        nodesPerAgent: 5,
-        vaultSeedNodeCount: 200,
-        perAgentTimeoutMs: 60_000,
-        bootTimeoutMs: 60_000,
-        settleAfterStormMs: 2_000,
-        outPath: null,
-        keepArtifacts: false,
-    }
-    let agents = defaults.agents
-    let nodesPerAgent = defaults.nodesPerAgent
-    let vaultSeedNodeCount = defaults.vaultSeedNodeCount
-    let perAgentTimeoutMs = defaults.perAgentTimeoutMs
-    let bootTimeoutMs = defaults.bootTimeoutMs
-    let settleAfterStormMs = defaults.settleAfterStormMs
-    let outPath = defaults.outPath
-    let keepArtifacts = defaults.keepArtifacts
-
-    const intArg = (raw: string | undefined, name: string): number => {
-        const n = Number.parseInt(raw ?? '', 10)
-        if (!Number.isInteger(n) || n < 0) throw new Error(`bad --${name}: ${raw}`)
-        return n
-    }
-
-    for (let i = 0; i < argv.length; i++) {
-        const a = argv[i]
-        switch (a) {
-            case '--agents': agents = intArg(argv[++i], 'agents'); break
-            case '--nodes-per-agent': nodesPerAgent = intArg(argv[++i], 'nodes-per-agent'); break
-            case '--vault-seed-nodes': vaultSeedNodeCount = intArg(argv[++i], 'vault-seed-nodes'); break
-            case '--per-agent-timeout-ms': perAgentTimeoutMs = intArg(argv[++i], 'per-agent-timeout-ms'); break
-            case '--boot-timeout-ms': bootTimeoutMs = intArg(argv[++i], 'boot-timeout-ms'); break
-            case '--settle-after-storm-ms': settleAfterStormMs = intArg(argv[++i], 'settle-after-storm-ms'); break
-            case '--out': outPath = argv[++i] ?? null; break
-            case '--keep-artifacts': keepArtifacts = true; break
-            case '--help':
-            case '-h':
-                process.stdout.write(
-                    'electron-main-storm.ts: profile Electron main CPU under an N-agent fake-agent storm.\n'
-                    + '  --agents N                    parallel fake-agents (default 5)\n'
-                    + '  --nodes-per-agent N           create_node actions per agent (default 5)\n'
-                    + '  --vault-seed-nodes N          seed-vault size (default 200)\n'
-                    + '  --per-agent-timeout-ms MS     per-agent completion deadline (default 60000)\n'
-                    + '  --boot-timeout-ms MS          how long to wait for app boot + .mcp.json (default 60000)\n'
-                    + '  --settle-after-storm-ms MS   keep profiling N ms after last agent exits (default 2000)\n'
-                    + '  --out PATH                    .cpuprofile path (default ~/.voicetree/reports/electron-main-storm-<ts>.cpuprofile)\n'
-                    + '  --keep-artifacts              keep temp vault + userData after the run\n',
-                )
-                process.exit(0)
-                break
-            default:
-                throw new Error(`unknown argument: ${a}`)
-        }
-    }
-    return {
-        agents, nodesPerAgent, vaultSeedNodeCount, perAgentTimeoutMs,
-        bootTimeoutMs, settleAfterStormMs, outPath, keepArtifacts,
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Repo + binary resolution
-// ---------------------------------------------------------------------------
-
-const __filename = fileURLToPath(import.meta.url)
-const __dirname = dirname(__filename)
-const require = createRequire(import.meta.url)
-
-// measures/perf -> measures -> packages -> repo root
-const REPO_ROOT = resolve(__dirname, '..', '..', '..')
-
-function resolveElectronBinary(): string {
-    // `require('electron')` returns the absolute path to the platform binary
-    // (e.g. .../node_modules/electron/dist/electron on Linux,
-    //       .../node_modules/electron/dist/Electron.app/Contents/MacOS/Electron on macOS).
-    // Resolve from webapp/ so we get the workspace-local copy electron-vite uses.
-    const electronModuleEntry = require.resolve('electron', {
-        paths: [join(REPO_ROOT, 'webapp'), REPO_ROOT],
-    })
-    const binary = require(electronModuleEntry) as unknown
-    if (typeof binary !== 'string') {
-        throw new Error(`require('electron') did not return a string path (got ${typeof binary})`)
-    }
-    if (!existsSync(binary)) {
-        throw new Error(
-            `electron binary missing at ${binary}\n`
-            + `The electron npm package was installed but its postinstall didn't download the\n`
-            + `platform binary. Run \`npm rebuild electron\` (or reinstall) on this machine.`,
-        )
-    }
-    return binary
-}
-
-function resolveMainBundleEntry(): string {
-    const entry = join(REPO_ROOT, 'webapp', 'dist-electron', 'main', 'index.js')
-    if (!existsSync(entry)) {
-        throw new Error(
-            `Built electron main bundle missing at ${entry}.\n`
-            + `Run \`npm --workspace webapp exec -- electron-vite build\` first.`,
-        )
-    }
-    return entry
-}
-
-function resolveFakeAgentEntrypoint(): { dir: string; entry: string } {
-    const dir = join(REPO_ROOT, 'tools', 'vt-fake-agent')
-    const entry = join(dir, 'src', 'index.ts')
-    if (!existsSync(entry)) throw new Error(`vt-fake-agent entrypoint not found at ${entry}`)
-    return { dir, entry }
-}
-
-function resolveTsxImportPath(): string {
-    return require.resolve('tsx')
-}
-
-// ---------------------------------------------------------------------------
-// Electron lifecycle
-// ---------------------------------------------------------------------------
-
-/**
- * Spawn the prebuilt electron app with `--inspect=0` and capture the inspect
- * port from stderr. Returns once Node has emitted the `Debugger listening` line.
- */
-async function spawnElectron(args: {
-    electronBinary: string
-    mainEntry: string
-    userDataDir: string
-    openFolder: string
-    bootTimeoutMs: number
-}): Promise<{ proc: ChildProcessWithoutNullStreams; inspectPort: number }> {
-    // Linux dev boxes (Onidel) run as root and have no X server; mirror the
-    // flag set used by webapp's e2e specs (no-sandbox + swiftshader + dev-shm).
-    // On macOS none of these are needed.
-    const linuxFlags = process.platform === 'linux'
-        ? ['--no-sandbox', '--disable-dev-shm-usage', '--use-gl=angle', '--use-angle=swiftshader', '--enable-unsafe-swiftshader']
-        : []
-    const proc = spawn(
-        args.electronBinary,
-        [
-            '--inspect=0',
-            ...linuxFlags,
-            args.mainEntry,
-            `--user-data-dir=${args.userDataDir}`,
-            // --open-folder bypasses the project-picker UI and tells main to
-            // open this vault directly on launch (see environment-config.ts).
-            '--open-folder', args.openFolder,
-        ],
-        {
-            env: {
-                ...process.env,
-                NODE_ENV: 'test',
-                VOICETREE_PERSIST_STATE: '1',
-            },
-            stdio: ['ignore', 'pipe', 'pipe'],
-        },
-    ) as ChildProcessWithoutNullStreams
-
-    const inspectPort = await new Promise<number>((resolveP, rejectP) => {
-        const timeout = setTimeout(
-            () => rejectP(new Error(`timed out waiting for --inspect port after ${args.bootTimeoutMs}ms`)),
-            args.bootTimeoutMs,
-        )
-        let stderrBuf = ''
-        proc.stderr.on('data', (chunk: Buffer) => {
-            const text = chunk.toString()
-            stderrBuf += text
-            // Echo electron stderr to our stderr so boot errors are visible.
-            process.stderr.write(`[electron] ${text}`)
-            const match = stderrBuf.match(/Debugger listening on ws:\/\/127\.0\.0\.1:(\d+)\//)
-            if (match) {
-                clearTimeout(timeout)
-                resolveP(Number.parseInt(match[1], 10))
-            }
-        })
-        proc.on('error', (err) => { clearTimeout(timeout); rejectP(err) })
-        proc.on('exit', (code, signal) => {
-            clearTimeout(timeout)
-            rejectP(new Error(`electron exited before inspect port appeared (code=${code} signal=${signal})`))
-        })
-    })
-
-    // Continue echoing electron stdout for visibility, but stop accumulating.
-    proc.stdout.on('data', (chunk: Buffer) => {
-        process.stderr.write(`[electron] ${chunk.toString()}`)
-    })
-    return { proc, inspectPort }
-}
-
-/**
- * Poll `<vault>/.mcp.json` for the voicetree MCP port. This file is written by
- * the electron app once its in-process MCP server has bound a port.
- */
-async function waitForMcpPort(vault: string, timeoutMs: number): Promise<number> {
-    const mcpPath = join(vault, '.mcp.json')
-    const deadline = Date.now() + timeoutMs
-    while (Date.now() < deadline) {
-        if (existsSync(mcpPath)) {
-            try {
-                const raw = readFileSync(mcpPath, 'utf8')
-                const cfg = JSON.parse(raw) as {
-                    mcpServers?: Record<string, { url?: string }>
-                }
-                const url = cfg.mcpServers?.voicetree?.url
-                if (url) {
-                    const m = url.match(/:(\d+)\/mcp$/)
-                    if (m) return Number.parseInt(m[1], 10)
-                }
-            } catch {
-                // file may be mid-write; retry
-            }
-        }
-        await new Promise((r) => setTimeout(r, 200))
-    }
-    throw new Error(`timed out waiting for ${mcpPath} after ${timeoutMs}ms`)
-}
-
-async function stopElectron(proc: ChildProcessWithoutNullStreams): Promise<void> {
-    if (proc.exitCode !== null || proc.signalCode !== null) return
-    proc.kill('SIGTERM')
-    await new Promise<void>((resolveP) => {
-        const force = setTimeout(() => { try { proc.kill('SIGKILL') } catch { /* */ } }, 5000)
-        proc.on('exit', () => { clearTimeout(force); resolveP() })
-    })
-}
-
-// ---------------------------------------------------------------------------
-// Fake-agent storm
-// ---------------------------------------------------------------------------
-
-function buildFakeAgentScript(nodesPerAgent: number): object {
-    const actions: object[] = []
-    for (let i = 0; i < nodesPerAgent; i++) {
-        actions.push({
-            type: 'create_node',
-            title: `Perf Node ${i}`,
-            summary: `Synthetic node ${i} from electron-main-storm.`,
-            content: `Node body ${i}.`,
-        })
-    }
-    actions.push({ type: 'exit', code: 0 })
-    return { actions }
-}
-
-interface AgentResult {
-    readonly terminalId: string
-    readonly spawnSuccess: boolean
-    readonly exitCode: number | null
-    readonly exitedAtMs: number | null
-    readonly errorMessage?: string
-}
-
-async function waitForExit(
-    terminalId: string,
-    exitedTerminals: Map<string, { code: number; atMs: number }>,
-    timeoutMs: number,
-): Promise<{ code: number; atMs: number } | null> {
-    const deadline = Date.now() + timeoutMs
-    while (Date.now() < deadline) {
-        const found = exitedTerminals.get(terminalId)
-        if (found) return found
-        const record = agentRuntime.getTerminalRecords().find((r) => r.terminalId === terminalId)
-        if (record?.status === 'exited') {
-            const entry = { code: 0, atMs: Date.now() }
-            exitedTerminals.set(terminalId, entry)
-            return entry
-        }
-        await new Promise((r) => setTimeout(r, 200))
-    }
-    return null
-}
-
-async function runStorm(args: {
-    mcpPort: number
-    vault: string
-    appSupport: string
-    agents: number
-    nodesPerAgent: number
-    perAgentTimeoutMs: number
-}): Promise<readonly AgentResult[]> {
-    const { dir: fakeAgentDir, entry: fakeAgentEntrypoint } = resolveFakeAgentEntrypoint()
-    const tsxImportPath = resolveTsxImportPath()
-
-    // graph-model is initialised by `startDaemon` in the daemon-only harness;
-    // here Electron owns the daemon, so init it locally for the in-process
-    // agentRuntime (`loadSettings` → `getSettingsPath` → `getConfig`).
-    initGraphModel({ appSupportPath: args.appSupport })
-
-    configureAgentRuntime({
-        env: {
-            getAppSupportPath: (): string => args.appSupport,
-            getMcpPort: (): number => args.mcpPort,
-        },
-    })
-
-    await agentRuntime.ensureTmuxAvailable()
-    await agentRuntime.ensureTmuxServer()
-
-    const script = buildFakeAgentScript(args.nodesPerAgent)
-    const agentPrompt = `### FAKE_AGENT_SCRIPT ###\n${JSON.stringify(script)}\n### END_FAKE_AGENT_SCRIPT ###`
-
-    const exitedTerminals = new Map<string, { code: number; atMs: number }>()
-    const onData = (_id: string, _data: string): void => { /* drop; profile is the artifact */ }
-    const onExit = (id: string, exitCode: number): void => {
-        if (!exitedTerminals.has(id)) exitedTerminals.set(id, { code: exitCode, atMs: Date.now() })
-    }
-
-    const launches: Promise<AgentResult>[] = []
-    for (let i = 0; i < args.agents; i++) {
-        const terminalId = `perf-agent-${i}` as TerminalId
-        const initialEnvVars: Record<string, string> = {
-            VOICETREE_TERMINAL_ID: terminalId,
-            VOICETREE_MCP_PORT: String(args.mcpPort),
-            VOICETREE_VAULT_PATH: args.vault,
-            TASK_NODE_PATH: `${args.vault}/${terminalId}-task.md`,
-            AGENT_PROMPT: agentPrompt,
-        }
-        const td: TerminalData = createTerminalData({
-            terminalId,
-            attachedToNodeId: args.vault,
-            terminalCount: i,
-            title: terminalId,
-            agentName: terminalId,
-            isHeadless: true,
-            initialEnvVars,
-            initialCommand: `${JSON.stringify(process.execPath)} --import ${JSON.stringify(tsxImportPath)} ${JSON.stringify(fakeAgentEntrypoint)}; exit`,
-            executeCommand: true,
-            initialSpawnDirectory: fakeAgentDir,
-        })
-
-        launches.push((async (): Promise<AgentResult> => {
-            const spawnRes = await agentRuntime.getTerminalManager().spawnTmuxBacked({
-                terminalData: td,
-                getToolsDirectory: () => fakeAgentDir,
-                onData,
-                onExit,
-            })
-            if (!spawnRes.success) {
-                return {
-                    terminalId,
-                    spawnSuccess: false,
-                    exitCode: -1,
-                    exitedAtMs: Date.now(),
-                    errorMessage: spawnRes.error ?? 'spawn failed',
-                }
-            }
-            const exit = await waitForExit(terminalId, exitedTerminals, args.perAgentTimeoutMs)
-            return {
-                terminalId,
-                spawnSuccess: true,
-                exitCode: exit?.code ?? null,
-                exitedAtMs: exit?.atMs ?? null,
-            }
-        })())
-    }
-
-    return Promise.all(launches)
-}
-
-// ---------------------------------------------------------------------------
-// Main
-// ---------------------------------------------------------------------------
+import { parseArgs } from './electron-main-storm/args.ts'
+import { spawnElectron, stopElectron, waitForMcpPort } from './electron-main-storm/electron.ts'
+import { resolveElectronBinary, resolveMainBundleEntry } from './electron-main-storm/paths.ts'
+import { runStorm } from './electron-main-storm/storm.ts'
 
 async function main(): Promise<void> {
     const args = parseArgs(process.argv.slice(2))
